chore(EnabledFor):renamed to enabledForProvider
