+++ src/Extensions/EnumerableExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using form_builder.EnabledFor;
 using form_builder.Enum;
 using form_builder.Providers.Address;
 using form_builder.Providers.Booking;
@@ -9,6 +8,7 @@ using form_builder.Providers.Lookup;
 using form_builder.Providers.Organisation;
 using form_builder.Providers.Street;
 using form_builder.Providers.Submit;
+using form_builder.Providers.EnabledFor;
 using form_builder.Providers.TemplatedEmailProvider;
 using form_builder.TagParsers.Formatters;
 
@@ -39,7 +39,8 @@ namespace form_builder.Extensions
 
         public static ITemplatedEmailProvider Get(this IEnumerable<ITemplatedEmailProvider> value, string providerName) =>
            value.Single(_ => _.ProviderName == providerName);
-        public static IEnabledFor Get(this IEnumerable<IEnabledFor> value, EEnabledFor providerName) =>
+
+        public static IEnabledForProvider Get(this IEnumerable<IEnabledForProvider> value, EEnabledFor providerName) =>
            value.Single(_ => _.Type.Equals(providerName));
     }
 }
