+++ src/Providers/EnabledFor/IEnabledForProvider.cs
@@ -1,9 +1,9 @@
 using form_builder.Enum;
 using form_builder.Models;
 
-namespace form_builder.EnabledFor
+namespace form_builder.Providers.EnabledFor
 {
-    public interface IEnabledFor
+    public interface IEnabledForProvider
     {
         EEnabledFor Type { get; }
         bool IsAvailable(EnabledForBase enabledFor);
